requirements.txt
mido==1.2.9
mingus==0.5.1
python-rtmidi==1.3.0
RPi.GPIO==0.6.5
